Migrate client tests into their own file
